松组合 release 1.0.0

+++ src/imu/navmech.cc
@@ -117,13 +117,13 @@ Eigen::MatrixXd MechTransferMat(const ImuData &pre_imu_data, const ImuData &curr
     static int scale_of_acce = config->get<int>("evaluate_imu_scale") == 0 ? 0 : 3;
     static int scale_of_gyro = scale_of_acce;
     static int rows = 15 + scale_of_acce + scale_of_gyro, cols = rows;
-    static int corr_time_of_gyro_bias = config->get<int>("corr_time_of_gyro_bias") * constant_hour;
-    static int corr_time_of_acce_bias = config->get<int>("corr_time_of_acce_bias") * constant_hour;
-    static int corr_time_of_gyro_scale = 0, corr_time_of_acce_scale = 0;
+    static double corr_time_of_gyro_bias = config->get<double>("corr_time_of_gyro_bias") * constant_hour;
+    static double corr_time_of_acce_bias = config->get<double>("corr_time_of_acce_bias") * constant_hour;
+    static double corr_time_of_gyro_scale = 0, corr_time_of_acce_scale = 0;
     if (scale_of_gyro == 3)
-        corr_time_of_gyro_scale = config->get<int>("corr_time_of_gyro_scale") * constant_hour;
+        corr_time_of_gyro_scale = config->get<double>("corr_time_of_gyro_scale") * constant_hour;
     if (scale_of_acce == 3)
-        corr_time_of_acce_scale = config->get<int>("corr_time_of_acce_scale") * constant_hour;
+        corr_time_of_acce_scale = config->get<double>("corr_time_of_acce_scale") * constant_hour;
 
     auto &pos = nav_info.pos_;
     auto &vel = nav_info.vel_;
